Handle empty tests_err and tests_ok directories

+++ tests.sh
@@ -52,6 +52,9 @@ run_tests_err() {
 	# Loop over all directories in tests_err/ (non-recursively)
 	for dir in tests_err/*/
 	do
+		# The wildcard doesn't expand if the directory is empty
+		[ -d "$dir" ] || continue
+
 		rm -f results/*
 		run_test_err $dir
 	done
@@ -118,6 +121,9 @@ run_tests_ok() {
 	# Loop over all directories in tests_ok/ (non-recursively)
 	for dir in tests_ok/*/
 	do
+		# The wildcard doesn't expand if the directory is empty
+		[ -d "$dir" ] || continue
+
 		rm -f results/*
 		run_test_ok $dir
 	done
